+++ client/ayon_substancedesigner/plugins/create/create_textures.py
@@ -1,18 +1,14 @@
 # -*- coding: utf-8 -*-
 """Creator plugin for creating textures."""
-from ayon_core.pipeline import CreatedInstance, Creator, CreatorError
+from ayon_core.pipeline import CreatorError
 from ayon_core.lib import BoolDef, EnumDef
 
-from ayon_substancedesigner.api.pipeline import (
-    get_instances,
-    set_instance,
-    set_instances,
-    remove_instance
-)
+from ayon_substancedesigner.api.pipeline import set_instance
 from ayon_substancedesigner.api.lib import get_current_graph_name
+from ayon_substancedesigner.api.plugin import TextureCreator
 
 
-class CreateTextures(Creator):
+class CreateTextures(TextureCreator):
     """Create a texture set."""
     identifier = "io.ayon.creators.substancedesigner.textureset"
     label = "Textures"
@@ -47,49 +43,12 @@ class CreateTextures(Creator):
             instance_data=instance.data_to_store()
         )
 
-    def collect_instances(self):
-        for instance in get_instances():
-            if (instance.get("creator_identifier") == self.identifier or
-                    instance.get("productType") == self.product_type):
-                self.create_instance_in_context_from_existing(instance)
-
-    def update_instances(self, update_list):
-        instance_data_by_id = {}
-        for instance, _changes in update_list:
-            # Persist the data
-            instance_id = instance.get("instance_id")
-            instance_data = instance.data_to_store()
-            instance_data_by_id[instance_id] = instance_data
-        set_instances(instance_data_by_id, update=True)
-
-    def remove_instances(self, instances):
-        for instance in instances:
-            remove_instance(instance["instance_id"])
-            self._remove_instance_from_context(instance)
-
-    # Helper methods (this might get moved into Creator class)
-    def create_instance_in_context(self, product_name, data):
-        instance = CreatedInstance(
-            self.product_type, product_name, data, self
-        )
-        self.create_context.creator_adds_instance(instance)
-        return instance
-
-    def create_instance_in_context_from_existing(self, data):
-        instance = CreatedInstance.from_existing(data, self)
-        self.create_context.creator_adds_instance(instance)
-        return instance
-
     def get_instance_attr_defs(self):
         return [
             BoolDef("review",
                     label="Review",
                     tooltip="Mark as reviewable",
                     default=True),
-            BoolDef("sbsar",
-                    label="Export Sbsar",
-                    tooltip="Export Sbsar along with textures",
-                    default=False),
             EnumDef("exportFileFormat",
                     items={
                         # TODO: Get available extensions from substance API
