+++ client/ayon_substancedesigner/plugins/publish/collect_textureset_images.py
@@ -38,9 +38,6 @@ class CollectTextureSet(pyblish.api.InstancePlugin):
             use_local_temp=True
         )
         instance.data["map_identifiers"] = map_identifiers
-        # if sbsar sets to True, it would enable to export sbsar
-        if instance.data["creator_attributes"].get("sbsar", False):
-            instance.data["families"].append("sbsar")
 
         for map_identifier in map_identifiers:
             self.create_image_instance(
