+++ client/ayon_substancedesigner/plugins/publish/extract_sbsar.py
@@ -14,7 +14,7 @@ class ExtractSbsar(publish.Extractor):
     hosts = ["substancedesigner"]
     families = ["sbsar"]
 
-    order = publish.Extractor.order - 0.099
+    order = publish.Extractor.order
 
     def process(self, instance):
         ctx = sd.getContext()
@@ -24,18 +24,17 @@ class ExtractSbsar(publish.Extractor):
         filename = os.path.basename(current_file)
         filename = filename.replace("sbs", "sbsar")
         staging_dir = self.staging_dir(instance)
-        sbsar_staging_dir = os.path.join(staging_dir, "sbsar")
         filepath = os.path.normpath(
-            os.path.join(sbsar_staging_dir, filename))
-
-        # export the graph with filepath
+            os.path.join(staging_dir, filename))
         exporter.exportSBSFileToSBSAR(current_file, filepath)
 
+        if "representations" not in instance.data:
+            instance.data["representations"] = []
         representation = {
             'name': 'sbsar',
             'ext': 'sbsar',
             'files': filename,
-            "stagingDir": sbsar_staging_dir,
+            "stagingDir": staging_dir,
         }
 
         instance.data["representations"].append(representation)
